kube-aws: VPC w/ Public and Private Subnets (NAT)
This change allows you to create a VPC with a public subnet and a
private subnet. The instances in the public subnet can receive inbound
traffic directly from the Internet, whereas the instances in the private
subnet can't. The instances in the public subnet can send outbound
traffic directly to the Internet, whereas the instances in the private
subnet can't. Instead, the instances in the private subnet can  access
the Internet by using a network address translation (NAT) gateway that
resides in the public subnet.

The controller node will be placed on the public subnet while all worker
nodes will be placed on the private subnet.